Update smac_input_readers.py
I've been using this project to merge parameter spaces of some SAT solvers for which I have PCS files. So thanks for writing the `read_pcs` and `merge_configuration_space` functions! However, I had no way to get a PCS file back out of this library, so I had to write this function to get the pcs file back out.